Auto-close for hamburger button!
Through use of an additional javascript event listener, we now
have the ability to open the hamburger menu and allow it to close
automatically close should a user on their desktop choose to widen
their browser window while the menu is still opened. Took quite a bit
of searching to get this one done, but we did it!